Refactor Dottore, Paziente, Tutore and Notifica
rest and service
