--- src/main/java/com/glucoseguardian/webbackend/application/service/dottore/AbstractDottoreService.java
@@ -1,8 +1,0 @@
-package com.glucoseguardian.webbackend.application.service.dottore;
-
-/**
- * This is an abstact class of DottoreService.
- * */
-public abstract class AbstractDottoreService {
-  public abstract DottoreServiceInterface getImplementation();
-}
