--- src/main/java/com/glucoseguardian/webbackend/application/service/notifica/AbstractNotificaService.java
@@ -1,10 +1,0 @@
-package com.glucoseguardian.webbackend.application.service.notifica;
-
-/**
- * This is an abstact class of NotificaService.
- */
-public abstract class AbstractNotificaService {
-
-  public abstract NotificaServiceInterface getImplementation();
-}
-
