--- src/main/java/com/glucoseguardian/webbackend/application/service/paziente/AbstractPazienteService.java
@@ -1,8 +1,0 @@
-package com.glucoseguardian.webbackend.application.service.paziente;
-
-/**
- * This is an abstact class of PazienteService.
- */
-public abstract class AbstractPazienteService {
-  public abstract PazienteServiceInterface getImplementation();
-}
