+++ src/main/java/com/glucoseguardian/webbackend/application/service/terapia/FinalTerapiaService.java
@@ -1,7 +1,5 @@
 package com.glucoseguardian.webbackend.application.service.terapia;
 
-import com.glucoseguardian.webbackend.application.service.paziente.PazienteServiceConcrete;
-import com.glucoseguardian.webbackend.application.service.paziente.PazienteServiceInterface;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
