--- src/main/java/com/glucoseguardian/webbackend/application/service/tutore/AbstractTutoreService.java
@@ -1,10 +1,0 @@
-package com.glucoseguardian.webbackend.application.service.tutore;
-
-
-/**
- * This is an abstact class of TutoreService.
- */
-public abstract class AbstractTutoreService {
-
-  public abstract TutoreServiceInterface getImplementation();
-}
