+++ src/main/java/com/glucoseguardian/webbackend/dottore/service/AbstractDottoreService.java
@@ -1,0 +1,8 @@
+package com.glucoseguardian.webbackend.dottore.service;
+
+/**
+ * This is an abstact class of DottoreService.
+ * */
+public abstract class AbstractDottoreService {
+  public abstract DottoreServiceInterface getImplementation();
+}
