+++ src/main/java/com/glucoseguardian/webbackend/notifica/service/AbstractNotificaService.java
@@ -1,0 +1,10 @@
+package com.glucoseguardian.webbackend.notifica.service;
+
+/**
+ * This is an abstact class of NotificaService.
+ */
+public abstract class AbstractNotificaService {
+
+  public abstract NotificaServiceInterface getImplementation();
+}
+
