+++ src/main/java/com/glucoseguardian/webbackend/paziente/service/AbstractPazienteService.java
@@ -1,0 +1,8 @@
+package com.glucoseguardian.webbackend.paziente.service;
+
+/**
+ * This is an abstact class of PazienteService.
+ */
+public abstract class AbstractPazienteService {
+  public abstract PazienteServiceInterface getImplementation();
+}
