+++ src/main/java/com/glucoseguardian/webbackend/tutore/service/AbstractTutoreService.java
@@ -1,0 +1,10 @@
+package com.glucoseguardian.webbackend.tutore.service;
+
+
+/**
+ * This is an abstact class of TutoreService.
+ */
+public abstract class AbstractTutoreService {
+
+  public abstract TutoreServiceInterface getImplementation();
+}
